+++ CHANGELOG.md
@@ -67,6 +67,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* [#851](https://github.com/allora-network/allora-chain/pull/851) Enforce recommended settings
+* [#835](https://github.com/allora-network/allora-chain/pull/835): Introduce the `x/scheduler` module to manage tasks scheduling
+
 ### Changed
 
 ### Deprecated
@@ -75,9 +78,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* [#871](https://github.com/allora-network/allora-chain/pull/871) Remove recover from Begin/EndBlockers, align mint/emissions
 * [#876](https://github.com/allora-network/allora-chain/pull/876) Fix jagged OOIF matrix for events containing value bundle
 
 ### Security
+
+* [#877](https://github.com/allora-network/allora-chain/pull/877) Upgrade Comet v0.38.19
 
 ### API Breaking Changes
 
@@ -89,7 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Released]
 
-# v0.12.4
+# v0.12.4
 
 ### Added
 
@@ -114,8 +120,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Changed
 
-
-# v0.12.3
+# v0.12.3
 
 ### Added
 
@@ -156,6 +161,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+
+[#846](https://github.com/allora-network/allora-chain/pull/846) Use v5 in mint module autocli txs
 
 ### Security